added updated tests for the battery and engine
