+++ test/test_battery/test_nubblin_battery.py
@@ -1,0 +1,17 @@
+import unittest
+from datetime import date
+
+from battery.NubbinBattery import NubbinBattery
+
+class TestNubblinBattery(unittest.TestCase):
+    def test_battery_should_be_serviced(self):
+        current_date = date.fromisoformat("2020-05-15")
+        last_service_date = date.fromisoformat("2016-01-25")
+        battery = NubbinBattery(current_date, last_service_date)
+        self.assertTrue(battery.needs_service())
+        
+    def test_battery_should_not_be_serviced(self):
+        current_date = date.fromisoformat("2020-05-15")
+        last_service_date = date.fromisoformat("2019-01-10")
+        battery = NubbinBattery(current_date, last_service_date)
+        self.assertFalse(battery.needs_service())
